+++ evmd/ante/evm_handler.go
@@ -37,10 +37,10 @@ func newLegacyCosmosAnteHandlerEip712(ctx sdk.Context, options HandlerOptions, e
 		txFeeChecker = evm.NewDynamicFeeChecker(ethCfg, &evmParams, &feemarketParams)
 	}
 	decorators := []sdk.AnteDecorator{
-		cosmos.RejectMessagesDecorator{},           // reject MsgEthereumTxs
-		cosmos.NewRejectStakingMessagesDecorator(), // reject staking messages
+		cosmos.RejectMessagesDecorator{}, // reject MsgEthereumTxs
 		// disable the Msg types that cannot be included on an authz.MsgExec msgs field
 		cosmos.NewAuthzLimiterDecorator(options.DisabledAuthzMsgs),
+		cosmos.NewRejectStakingMessagesDecorator(), // reject staking messages
 		authante.NewSetUpContextDecorator(),
 		authante.NewValidateBasicDecorator(),
 		authante.NewTxTimeoutHeightDecorator(),