+++ evmd/ante/handler_options.go
@@ -183,10 +183,10 @@ func newCosmosAnteHandler(ctx sdk.Context, options HandlerOptions, extra ...sdk.
 		txFeeChecker = evm.NewDynamicFeeChecker(ethCfg, &evmParams, &feemarketParams)
 	}
 	decorators := []sdk.AnteDecorator{
-		cosmos.RejectMessagesDecorator{},           // reject MsgEthereumTxs
-		cosmos.NewRejectStakingMessagesDecorator(), // reject staking messages
+		cosmos.RejectMessagesDecorator{}, // reject MsgEthereumTxs
 		// disable the Msg types that cannot be included on an authz.MsgExec msgs field
 		cosmos.NewAuthzLimiterDecorator(options.DisabledAuthzMsgs),
+		cosmos.NewRejectStakingMessagesDecorator(), // reject staking messages
 		ante.NewSetUpContextDecorator(),
 		ante.NewExtensionOptionsDecorator(options.ExtensionOptionChecker),
 		ante.NewValidateBasicDecorator(),